incremented module version to 1.1.0. update tiapp.xml on testbed.
