+++ MODULE_CHANGELOG.md
@@ -1,0 +1,7 @@
+# Change Log
+All notable changes to the Branch Titanium SDK will be documented here.
+
+## [v1.1.0] - 2016-05-31
+### New Functionality
+- **getCredits**: gets credits from default bucket
+- **getCreditsForBucket()**: gets credits from a specific bucket;
