+++ android/manifest
@@ -2,7 +2,7 @@
 # this is your module manifest and used by Titanium
 # during compilation, packaging, distribution, etc.
 #
-version: 1.0.0
+version: 1.1.0
 apiversion: 2
 architectures: armeabi armeabi-v7a x86
 description: Titanium-Deferred-Deep-Linking-SDK
