--- testbed/modules/android/io.branch.sdk/1.0.0/manifest
@@ -1,18 +1,0 @@
-#
-# this is your module manifest and used by Titanium
-# during compilation, packaging, distribution, etc.
-#
-version: 1.0.0
-apiversion: 2
-architectures: armeabi armeabi-v7a x86
-description: Titanium-Deferred-Deep-Linking-SDK
-author: Branch Metrics
-license: Specify your license
-copyright: Copyright (c) 2015 by Your Company
-
-# these should not be edited
-name: Titanium-Deferred-Deep-Linking-SDK
-moduleid: io.branch.sdk
-guid: df14a182-464d-4940-bc1d-ae84730366a8
-platform: android
-minsdk: 5.0.2.GA
