--- tests/modules/android/io.branch.sdk/1.0.0/documentation/index.html
@@ -1,22 +1,0 @@
-<h1>Titanium-Deferred-Deep-Linking-SDK Module</h1>
-<h2>Description</h2>
-<p>TODO: Enter your module description here</p>
-<h2>Accessing the Titanium-Deferred-Deep-Linking-SDK Module</h2>
-<p>To access this module from JavaScript, you would do the following:</p>
-<pre><code>var titanium_deferred_deep_linking_sdk = require("io.branch.sdk");
-</code></pre>
-<p>The titanium_deferred_deep_linking_sdk variable is a reference to the Module object.</p>
-<h2>Reference</h2>
-<p>TODO: If your module has an API, you should document
-the reference here.</p>
-<h3>titanium_deferred_deep_linking_sdk.function</h3>
-<p>TODO: This is an example of a module function.</p>
-<h3>titanium_deferred_deep_linking_sdk.property</h3>
-<p>TODO: This is an example of a module property.</p>
-<h2>Usage</h2>
-<p>TODO: Enter your usage example here</p>
-<h2>Author</h2>
-<p>TODO: Enter your author name, email and other contact
-details you want to share here.</p>
-<h2>License</h2>
-<p>TODO: Enter your license/legal information here.</p>
