+++ tests/modules/android/io.branch.sdk/1.1.0/LICENSE
@@ -1,0 +1,9 @@
+All components of this product are Copyright (c) 2014 Pawprint Labs, Inc. (d/b/a Branch Metrics).  All rights reserved.
+
+Subject to the terms of this notice, Branch Metrics grants you a nonexclusive, nontransferable license, without the right to sublicense, to (a) install and execute one copy of these files on any number of workstations owned or controlled by you and (b) distribute verbatim copies of these files to third parties.  As a condition to the foregoing grant, you must provide this notice along with each copy you distribute and you must not remove, alter, or obscure this notice. All other use, reproduction, modification, distribution, or other exploitation of these files is strictly prohibited, except as may be set forth in a separate written license agreement between you and Branch Metrics. The terms of any such license agreement will control over this notice.  The license stated above will be automatically terminated and revoked if you exceed its scope or violate any of the terms of this notice.
+
+This License does not grant permission to use the trade names, trademarks, service marks, or product names of Branch Metrics, except as required for reasonable and customary use of our files and reproducing the content of this notice.  You may not mark or brand our files with any trade name, trademarks, service marks, or product names other than the original brand (if any) provided by Branch Metrics, except as may be set forth in a separate written license agreement between you and Branch Metrics.
+
+You may not remove or hide any existing mark or brand from the files as provided by Branch Metrics.
+
+Unless otherwise expressly agreed by Branch Metrics in a separate written license agreement, these files are provided AS IS, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF  MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT  SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE. As a condition to your use of these files, you are solely responsible  for such use. Branch Metrics will have no liability to you for direct, indirect, consequential, incidental, special, or punitive damages or for lost profits or data.
